chore/server: remove function out of routing file
